entry: implement reference counting for the entry code
In order to allow for safe module unmapping.

Grab a reference to THIS_MODULE and drop it only
when, after replacing the IDT again, that counter
becomes zero.

Take special care of forks and exits.

Note: currently the patch_entry_drain_start() will send
fake signals by itself. This is for testing only and will
be dropped in the future as KGraft does this already on
transitions.

Signed-off-by: Nicolai Stange <nstange@suse.de>